Fix warnings from ICC
darcs-hash:20060329001450-ac50b-9c3fe4f7ca14a077997c421616c07dde03b48c6d.gz

+++ wildcard.c
@@ -605,7 +605,7 @@ int wildcard_expand( const wchar_t *wc,
 		/*
 		  The maximum length of a file element
 		*/
-		size_t ln=MAX_FILE_LENGTH;
+		long ln=MAX_FILE_LENGTH;
 		char * narrow_dir_string = wcs2str( dir_string );
 		
 		if( narrow_dir_string )
